refactor(cache): rename files
